Refactor clean architecture, save api response
