+++ main.go
@@ -3,13 +3,13 @@ package main
 import (
 	"log"
 	"net/http"
-	"os"
 
 	"github.com/s1nuh3/academy-go-q32021/config"
 	"github.com/s1nuh3/academy-go-q32021/controller"
+	"github.com/s1nuh3/academy-go-q32021/repository"
 	"github.com/s1nuh3/academy-go-q32021/routes"
 	"github.com/s1nuh3/academy-go-q32021/service/clientAPI"
-	"github.com/s1nuh3/academy-go-q32021/service/storage"
+	"github.com/s1nuh3/academy-go-q32021/service/user"
 	"github.com/s1nuh3/academy-go-q32021/usecase"
 )
 
@@ -18,28 +18,19 @@ func main() {
 	cfg := config.ReadConfig()
 
 	log.Println("Starting")
+	file := repository.New(cfg.Csv.Path + cfg.Csv.Name)
+	userService := user.New(file)
+	uc := usecase.NewUser(userService)
+	c := controller.New(uc)
 
-	repo := storage.New(cfg.Csv.Path + cfg.Csv.Name)
-	u := usecase.New(repo)
-	c := controller.New(u)
-
-	client := clientAPI.New(cfg)
-	iu := usecase.NewConsumeAPI(client)
+	client := clientAPI.New(cfg, file)
+	iu := usecase.NewConsumeAPI(client, userService)
 	ci := controller.NewImportCtrl(iu)
 
 	r := routes.New(c, ci)
-
 	port := cfg.Server.Port
 	http.HandleFunc("/", r.ServeHTTP)
 
 	log.Println("App running.. on port ", port)
-	err := http.ListenAndServe(port, nil)
-	errHandler(err)
-}
-
-func errHandler(e error) {
-	if e != nil {
-		log.Println(e)
-		os.Exit(1)
-	}
+	log.Fatal(http.ListenAndServe(port, nil))
 }
